+++ fmpr/exp.c
@@ -59,7 +59,7 @@ fmpr_expm1(fmpr_t y, const fmpr_t x, slong prec, fmpr_rnd_t rnd)
         else if (fmpr_is_pos_inf(x))
             fmpr_pos_inf(y);
         else if (fmpr_is_neg_inf(x))
-            fmpr_set_si(y, -1L);
+            fmpr_set_si(y, -WORD(1));
         else
             fmpr_nan(y);
